Single-cell Annotation Matrix

Single-cell Annotation Matrix

Single-cell Annotation Matrix

Details

Base class for matrix-like data with rows aligned to the observations or features of an SCGroup.

Super classes

tiledbsc::TileDBArray -> tiledbsc::AnnotationArray -> AnnotationMatrix

Public fields

uri

URI of the TileDB array

verbose

Print status messages

Methods

Public methods

AnnotationMatrix$from_matrix()

AnnotationMatrix$to_matrix()

AnnotationMatrix$clone()

Inherited methods

tiledbsc::TileDBArray$add_metadata()

tiledbsc::TileDBArray$array_exists()

tiledbsc::TileDBArray$attributes()

tiledbsc::TileDBArray$attrnames()

tiledbsc::TileDBArray$class()

tiledbsc::TileDBArray$dimensions()

tiledbsc::TileDBArray$dimnames()

tiledbsc::TileDBArray$fragment_count()

tiledbsc::TileDBArray$get_metadata()

tiledbsc::TileDBArray$initialize()

tiledbsc::TileDBArray$print()

tiledbsc::TileDBArray$schema()

tiledbsc::TileDBArray$tiledb_array()

Method from_matrix()

Ingest annotation matrix

Usage

AnnotationMatrix$from_matrix(x, index_col)

Arguments

x

a matrix with charater vectors used for row/column names

index_col

Name to use for the TileDB array's dimension that will contain the matrix row names.

Method to_matrix()

Retrieve the annotation data from TileDB

Usage

AnnotationMatrix$to_matrix()

Returns

A matrix

Method clone()

The objects of this class are cloneable with this method.

Usage

AnnotationMatrix$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.