Base class for interacting with TileDB groups

Base class for interacting with TileDB groups

Base class for interacting with TileDB groups

Value

A named list of arrays. The array object.

Public fields

uri

The URI of the TileDB group

arrays

Named list of arrays in the group

verbose

Whether to print verbose output

Methods

Public methods

TileDBGroup$new()

TileDBGroup$class()

TileDBGroup$print()

TileDBGroup$group_exists()

TileDBGroup$list_objects()

TileDBGroup$list_object_uris()

TileDBGroup$get_arrays()

TileDBGroup$get_array()

TileDBGroup$get_metadata()

TileDBGroup$add_metadata()

TileDBGroup$clone()

Method new()

Create a new TileDBGroup object.

Usage

TileDBGroup$new(uri, verbose = TRUE)

Arguments

uri

TileDB array URI

verbose

Print status messages

Method class()

Print the name of the R6 class.

Usage

TileDBGroup$class()

Method print()

Print summary of the group.

Usage

TileDBGroup$print()

Method group_exists()

Check if the group exists.

Usage

TileDBGroup$group_exists()

Returns

TRUE if the group exists, FALSE otherwise.

Method list_objects()

List the TileDB objects within the group.

Usage

TileDBGroup$list_objects(type = NULL)

Arguments

type

The type of object to list, either "ARRAY", or "GROUP". By default all object types are listed.

Returns

A data.frame with columns URI and TYPE.

Method list_object_uris()

List URIs for TileDB objects within the group.

Usage

TileDBGroup$list_object_uris(type = NULL, prefix = NULL)

Arguments

type

The type of object to list, either "ARRAY", or "GROUP". By default all object types are listed.

prefix

Filter URIs whose basename contain an optional prefix.

Returns

A character vector of object URIs with names corresponding to the basename of the object.

Method get_arrays()

Retrieve arrays within the group that meet the specified criteria.

Usage

TileDBGroup$get_arrays(prefix = NULL)

Arguments

prefix

String prefix to filter the array names.

Method get_array()

Retrieve an array within the group.

Usage

TileDBGroup$get_array(name)

Arguments

name

The name of the array to retrieve.

Method get_metadata()

Retrieve metadata from the TileDB group.

Usage

TileDBGroup$get_metadata(key = NULL, prefix = NULL)

Arguments

key

The name of the metadata attribute to retrieve.

prefix

Filter metadata using an optional prefix. Ignored if key is not NULL.

Returns

A list of metadata values.

Method add_metadata()

Add list of metadata to the TileDB group.

Usage

TileDBGroup$add_metadata(metadata, prefix = "")

Arguments

metadata

Named list of metadata to add.

prefix

Optional prefix to add to the metadata attribute names.

Returns

NULL

Method clone()

The objects of this class are cloneable with this method.

Usage

TileDBGroup$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.